Invalidate cached crowd user if token changes
Normally a crowd user is cached in devise for a short period of time
(usually 10 minutes) so that we aren't constantly wasting time talking
to crowd for every request.  This commit caches the crowd.token_key as
well, so that if the user logs out in another site or logs in as someone
else on another site, they will be logged out immediately in devise.

If the token is invalidated by some other means (admin invalidates the
crowd token, or the user logs in on another computer) they may still
appeared logged in to devise until the next crowd authorization.